Sync Windows/ESENT changes to Exchange/ESE
Sync Windows/ESENT changes through commit 4a0e169cb9793.

[OS:4a0e169cb9793]

[Substrate:241509805d387723bee5196aab52decaf41a2a69]

+++ dev/ese/src/ese/revertsnapshot.cxx
@@ -4773,7 +4773,7 @@ ERR CRBSDatabaseRevertContext::ErrRBSInitRootPageDeleteState( const LONG lRBSGen
     QWORD cbOffset      = 0;
     QWORD cbRemaining   = cbSize;
 
-    pbread = (BYTE*)PvOSMemoryPageAlloc( cbSize, NULL );
+    pbread = (BYTE*)PvOSMemoryPageAlloc( (size_t)cbSize, NULL );
     Alloc( pbread );
 
     while ( cbRemaining > 0 )
